third_party/blink/web_tests/fast/forms/select/select-show-picker.html

<!DOCTYPE html>
<meta name=fuzzy content="maxDifference=0-3; totalPixels=0-1000">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>

<select id="select">
  <option>1</option>
</select>
<script>
promise_test(async (t) => {
  await test_driver.bless("show select picker");
  select.showPicker();
  assert_true(internals.isSelectPopupVisible(select), "Picker is opened.");
}, "Select: showPicker() should show popup");
</script>
